the_lab_3/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
psycopg2-binary
python-dotenv
requests
sqlalchemy
passlib
python-jose